requirements.txt
click
lmdb==1.2.1
natsort==7.1.1
nltk==3.6.2
numpy==1.21.6
opencv-python==4.5.1.48
protobuf==3.19.6
pyyaml
scikit-image==0.16.2
scipy==1.7.3
tensorboard==2.9.0
tensorflow==2.9.2
keras==2.9.0
tensorflow-estimator==2.9.0
lightgbm==3.3.2
transformers==4.19.1
pandas==1.4.3
wheel
